refactor(codec-image): extract landscape renderer from pipeline/decoder.ts (closes #325)
Per the AI-shape audit ([docs/research/2026-05-13-ai-shape-audit.md], merged
via #328), pipeline/decoder.ts (602 lines) was conflating four distinct
concerns: placeholder landscape rendering with magic-salt constants, scalar
field operations, PNG byte assembly, and the eventual M1B frozen-decoder
bridge entry point. The procedural landscape work was the strongest
AI-shape smell — it had 7+ magic-number constants embedded inline and
shared no cohesion with the field/PNG plumbing it sat next to.

This refactor splits the file along its responsibility seams:

1. `pipeline/internal-math.ts` (new, 89L) — shared math helpers
   (`clamp01`, `clampInt`, `clampByte`, `lerp`, `smoothstep`, `hash01`,
   `sampleNoise`, `mixColor`, `scaleColor`, `sampleField`) used by both
   the orchestrator and the landscape renderer. No duplication.

2. `pipeline/landscape-renderer.ts` (new, 321L) — placeholder landscape
   rendering with `SceneMode` / `SceneProfile` types, the four
   `FIELD_SALTS` magic constants hoisted to a named export with doc
   string, `buildSceneProfile`, `buildPalette`, `renderSky`,
   `renderTerrain`, private `terrainShade`, and the reference-landscape
   Python bridge (`tryDecodeReferenceLandscape` + `selectReferenceImage`).
   Self-contained for the landscape concern.

3. `pipeline/decoder.ts` (shrunk 602 → 309L) — now a thin orchestrator
   that builds scalar fields from latent tokens via the named
   `FIELD_SALTS`, runs the landscape renderer for pixel synthesis, and
   encodes the result as PNG. The `loadLlamagenDecoderBridge` wiring
   under `../decoders/llamagen.ts` is what eventually replaces the
   placeholder path here; M1B is gated on #283 per-candidate audits
   (PR #336 just landed the first per-candidate audit deliverable for
   VQGAN-class).

What did NOT change:

- Public API of pipeline/decoder.ts — `decodeLatentsToRaster` keeps the
  same signature including the `onChunk` M1B seam. Only one caller
  (`packages/codec-image/src/codec.ts:19`) and it's untouched.
- Pixel output bytes — the refactor moves code around but the actual
  computation per pixel is character-identical, so existing tests pass
  byte-for-byte.
- Magic salt values — `0x1a2b3c4d` etc. are now `FIELD_SALTS.elevation`
  etc. but the values are unchanged.

Validation:

- `pnpm --filter @wittgenstein/codec-image typecheck` clean.
- `pnpm --filter @wittgenstein/codec-image test` — all 45 tests pass.
- `pnpm --filter @wittgenstein/codec-image lint` — 0 errors, 0 warnings.
- `pnpm test:golden` — codec-audio parity + codec-sensor golden suites
  green (codec-image's parity is part of its main test suite).

decoder.ts shrinks from 602 to 309 lines, well under the audit's
"under 400 lines" target. Total LOC across the three files is slightly
higher than the original (719 vs 602) due to file headers, imports,
and the FIELD_SALTS doc strings — that's the cost of the seam, and
it's a good trade.

Refs:
- Closes #325.
- Audit source: docs/research/2026-05-13-ai-shape-audit.md (merged via #328).
- Related M1B work: #283 per-candidate audits (PR #336 just merged for
  VQGAN-class Gates A+B), #70 M1B umbrella.
